[quant][refactor] Move some util functions from torch/quantization/fx/utils.py to torch/quantization/utils.py
Summary:

Test Plan:

Reviewers:

Subscribers:

Tasks:

Tags:

[ghstack-poisoned]
