+++ torch/quantization/fx/quantization_patterns.py
@@ -13,20 +13,24 @@ from ..quantization_mappings import (
     get_static_quant_module_class,
     get_quantized_operator,
 )
+from ..utils import (
+    get_swapped_custom_module_class,
+    activation_is_statically_quantized,
+    weight_is_statically_quantized,
+    weight_dtype,
+    get_qconfig_dtypes,
+)
+
 from .pattern_utils import (
     register_quant_pattern,
     mark_input_output_not_observed,
 )
+
 from .utils import (
     _parent_name,
     quantize_node,
     get_per_tensor_qparams,
-    get_swapped_custom_module_class,
-    activation_is_statically_quantized,
-    weight_is_quantized,
-    weight_dtype,
     get_linear_prepack_op_for_dtype,
-    get_qconfig_dtypes,
 )
 
 from abc import ABC, abstractmethod
@@ -348,7 +352,7 @@ class LinearReLUQuantizeHandler(QuantizeHandler):
                 quantized_input_idxs = []
                 if activation_statically_quantized:
                     quantized_input_idxs.append(0)
-                if weight_is_quantized(qconfig):
+                if weight_is_statically_quantized(qconfig):
                     quantized_input_idxs.append(1)
                 args = load_arg(quantized=quantized_input_idxs)(self.linear_node.args)
                 args = load_arg(quantized=False)(self.linear_node.args)
@@ -369,7 +373,7 @@ class LinearReLUQuantizeHandler(QuantizeHandler):
             else:  # non-debug option
                 # linear args
                 # (x, weight, bias, ...)
-                weight_quantized = weight_is_quantized(qconfig)
+                weight_quantized = weight_is_statically_quantized(qconfig)
                 linear_weight = load_arg(quantized=weight_quantized)(self.linear_node.args[1])
 
                 # get other arguments
